Time: 52 ms (90.75%), Space: 61.8 MB (19.71%) - LeetHub

+++ 0011-container-with-most-water/0011-container-with-most-water.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int maxArea(vector<int>& height) {
-        long long max_val = 0 , min_val=0;
+        int max_val = 0 , min_val=0;
         int left = 0;
         int right = height.size()-1;
 
@@ -10,7 +10,7 @@ public:
             if (height[left] < height[right])  min_val = height[left];
             else min_val = height[right];
 
-            long long area = dis* min_val;
+            int area = dis* min_val;
 
             if(max_val < area) max_val = area;
